requirements.txt
pytest
pytest-json-report
GitPython
requests